// src/types/i-user.ts
export default interface IUser {
    id: number;
    userName: string;
    userRoles: string[];
    email: string;
    phoneNumber: string;
    firstName: string;
    lastName: string;
    fatherName: string;
    gender: string;
    dateOfBirth: string;
    country: string;
    city: string;
    group: string;
    photoUrl: string;
}